Separate debug function into another module.
